era 9/huggingface_deploy/requirements.txt
torch==2.0.1
torchvision==0.15.2
gradio==3.50.2
pillow==10.0.0 